+++ addOns/importurls/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Changed
 - Update minimum ZAP version to 2.8.0.
+- Add import menu to (new) top level Import menu instead of Tools menu.
 
 ## [6] - 2019-06-07